backend/requirements.txt
django
django-import-export
gunicorn
psycopg2
psycopg2-binary
python-dotenv
